backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
psycopg2-binary
python-dotenv
pyjwt
passlib[bcrypt]
